Revert "implement code review feedback"
This reverts commit 269bb756afcdd02a019458f06f0fb78d572cf2d0.

+++ lib/pavilion/lockfile.py
@@ -15,7 +15,6 @@ from time import sleep
 
 from pavilion import output
 from pavilion import utils
-from pavilion.errors import LockfileError
 
 # Expires after a silly long time.
 NEVER = 10**10
@@ -362,23 +361,10 @@ class FuzzyLock:
 
         :return: True if lock has been successfully acquired; False if timed out
         """
+        self._lock_dir.mkdir(exist_ok=True)
 
-        try:
-            self._lock_dir.mkdir(exist_ok=True)
-
-            # Declare intent to take lock
-            self._lockfile.touch(exist_ok=False)
-        except OSError:
-            msg = f"Unable to create lockfile {self._lockfile}."
-
-            if not self._lock_dir.exists():
-                # This is probably the fault of the caller
-                msg += " Lock directory ({self._lock_dir}) does not exist."
-            elif self._lockfile.exists():
-                # If this happens, something went very wrong. This is likely dead code.
-                msg += " Lockfile ({self._lock_dir}) already exists."
-            
-            raise LockFileError(msg)
+        # Declare intent to take lock
+        self._lockfile.touch(exist_ok=False)
 
         first = False
         start = time.time()
@@ -403,14 +389,13 @@ class FuzzyLock:
         del self._mtimes[self._lockfile]
 
         # Remove lock directory once it's empty
-        try:
-            # This will simply fail if the directory isn't empty,
-            # which is what we want.
-            self._lock_dir.rmdir()
-        except OSError:
-            # Another thread or process has removed the directory,
-            # or the directory is not empty; can be safely ignored
-            pass
+        if len(list(self._lock_dir.iterdir())) == 0:
+            try:
+                self._lock_dir.rmdir()
+            except FileNotFoundError:
+                # Another thread or process has removed the directory;
+                # can be safely ignored
+                pass
 
     def _get_earliest(self) -> Path:
         """Get the path to the lockfile that was created first.
@@ -418,8 +403,9 @@ class FuzzyLock:
         :return: Path object to whichever lockfile was created first"""
 
         lfiles = list(self._lock_dir.iterdir())
+        dict_keys = list(self._mtimes.keys())
 
-        for lockfile in list(self._mtimes.keys()):
+        for lockfile in dict_keys:
             if lockfile not in lfiles:
                 try:
                     del self._mtimes[lockfile]
